Rename variables and add `related_headword` and `dewk_main_form_on_variant` to output.
The renamed variables are `Redewendung` to `headword` and `Hauptform` to `main_form`.

Add code to check assigned canonical form vs Wiktionary main form.

Add run_wikwork.py as example code for obtaining the Wiktionary wikitext which can
be used to create an input file to `post_process.py` to get the Wiktionary main
form.

The `main_form` assigned was also changed for a number of headwords.