Move `VERB_FORMS` from global constant to parameter of `count_regexes`.
The purpose for this is to reduce the likelihood that `counts_regexes.py`
needs to be modified to process new idioms.

A few additional placeholders were added to the txt file that now contains
the placeholders and their replacement in the transition. These are not
needed to process the idioms that are headwords, but there is no harm in
including.

For the same reason `VERB_FORMS` was changed, we added more 'SEIN..'
placeholders so that all cases and genders were covered. Again, not
all of these are necessary to process the idioms that are headwords.

Since the user can now pass an arbitrary `verb_form` dict, we added a
check that all the keys use the convention we have been using and
consist of all capital letters and underscores.

We no longer count 'ß' as a capital letter when checking whether a regex
placeholder is all capital letters or underscores, as this is not a
capital letter. We removed the placeholder 'SCHIEßEN' that was present
in the previous `VERB_FORMS`. This has no effect on the output since
the input file was already using 'SCHIEẞEN', which also existed in
`VERB_FORMS`.